Use lazily-evaluated path
Summary:
1. Add LazyPath to fvcore
2. Allow using a path-like object in detectron2

Facebook
Sometimes certain non-trivial logic is needed before determining the path of a dataset, e.g.:
1. generate the path on-the-fly
1. benchmark each data sources and choose the best
1. preload memcache before using the dataset
1. cache the whole dataset locally
...

So it's better to determine the path when the dataset is used, not when detectron2 is imported.
Using a LazyPath can address this issue.
Currently it can print a message of the selected path when the dataset is used.

Reviewed By: rbgirshick

Differential Revision: D20052469

fbshipit-source-id: 2e958d28a4a13438038a6a468ed4b23b4bce93b0