Migrate fvcore tests off sandcastle gpu tier
Summary:
This contbuild is using the Sandcastle GPU tier which is no longer supported and ties up valuable resources needed by AI/ML Infra. Thus we would like this job to convert to utilize Remote Execution via tpx and the GPU worker there for far more efficiency and expected better performance.

Some of the unit tests focus on io which doesn't need a GPU and can be split off.

Reviewed By: ppwwyyxx

Differential Revision: D26517033

fbshipit-source-id: b0a5cf848bb6149a28ea77a25ceb0d13c445c564