Updated flop counting for linear layers to account for changed operator call from aten::addmm to aten::linear.
Summary: `nn.Linear` in Pytorch at version 1.9.0a0fb calls `aten::linear` instead of `aten::addmm` or `aten::matmul`, which breaks the flop counting. This adds a `aten::linear` flop and activation handle, adds it to the defaults for `FlopCountAnalysis` and `ActivationCountAnalysis`, and updates the tests to check for which operator is called and then tests against that operator.

Reviewed By: bxiong1202, ppwwyyxx

Differential Revision: D26573978

fbshipit-source-id: 94f9ff8b08260bb2370563dc73497e4b0448be18